+++ tests/Support/Rule/RuleWithCustomHandler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Validator\Tests\Support;
+namespace Yiisoft\Validator\Tests\Support\Rule;
 
 use Yiisoft\Validator\RuleInterface;
